fix: remaining errors

+++ adapter/in/shared/model.go
@@ -3,7 +3,7 @@ package shared
 const MoneyTransferTaskQueueName = "TRANSFER_MONEY_TASK_QUEUE"
 
 type PaymentDetails struct {
-	UserId      string `json:"userId"`
-	RecipientId string `json:"recipientId"`
-	Amount      int    `json:"amount"`
+	UserId      string  `json:"userId"`
+	RecipientId string  `json:"recipientId"`
+	Amount      float64 `json:"amount"`
 }
